frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.33.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
streamlit-quill = "^0.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
coverage = "^7.5.0"
pytest-cov = "^5.0.0"
ruff = "^0.4.3"
mypy = "^1.10.0"
pytest-mock = "^3.14.0"
pytest-httpx = "^0.30.0"
pytest-asyncio = "^0.23.7"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"


[tool.ruff]
exclude = [
    ".ruff_cache"
]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
    "F", # Flake8
    "B", # Black    
    "E", # error
    "W", # warning
    "I", # isort
]
ignore = []
fixable = ["ALL"]


[tool.ruff.lint.per-file-ignores]
"!app/**.py" = ["F", "B", "E", "W", "I"]
"__init__.py" = ["F", "B", "E", "W", "I"]


[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
# サードパーティの[import-untyped]を無視する
module = [
    'requests/*',
    'httpx/*',    
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
